# server/docs/swagger.yaml
definitions:
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        description: 业务状态码，0 表示成功
        type: integer
      data:
        description: 响应数据载荷
      msg:
        description: 提示信息
        type: string
    type: object
...
  system.SysDictionaryDetail:
    properties:
      ID:
        description: 主键ID
        type: integer
      children:
        description: 子字典详情
        items:
          $ref: '#/definitions/system.SysDictionaryDetail'
        type: array
      createdAt:
        description: 创建时间
        type: string
      disabled:
        description: 禁用状态，根据status字段动态计算
        type: boolean
      extend:
        description: 扩展值
        type: string
      label:
        description: 展示值
        type: string
      level:
        description: 层级深度，从0开始
        type: integer
      parentID:
        description: 父级字典详情ID
        type: integer
      path:
        description: 层级路径，如 "1,2,3"
        type: string
      sort:
        description: 排序标记
        type: integer
      status:
        description: 启用状态
        type: boolean
      sysDictionaryID:
        description: 关联标记
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      value:
        description: 字典值
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.9.2
paths:
...
  /sysDictionaryDetail/createSysDictionaryDetail:
    post:
      consumes:
      - application/json
      parameters:
      - description: SysDictionaryDetail模型
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysDictionaryDetail'
      produces:
      - application/json
      responses:
        "200":
          description: 创建SysDictionaryDetail
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建SysDictionaryDetail
      tags:
      - SysDictionaryDetail
  /sysDictionaryDetail/deleteSysDictionaryDetail:
    delete:
      consumes:
      - application/json
      parameters:
      - description: SysDictionaryDetail模型
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysDictionaryDetail'
      produces:
      - application/json
      responses:
        "200":
          description: 删除SysDictionaryDetail
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除SysDictionaryDetail
      tags:
      - SysDictionaryDetail
  /sysDictionaryDetail/findSysDictionaryDetail:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 禁用状态，根据status字段动态计算
        in: query
        name: disabled
        type: boolean
      - description: 扩展值
        in: query
        name: extend
        type: string
      - description: 展示值
        in: query
        name: label
        type: string
      - description: 层级深度，从0开始
        in: query
        name: level
        type: integer
      - description: 父级字典详情ID
        in: query
        name: parentID
        type: integer
      - description: 层级路径，如 "1,2,3"
        in: query
        name: path
        type: string
      - description: 排序标记
        in: query
        name: sort
        type: integer
      - description: 启用状态
        in: query
        name: status
        type: boolean
      - description: 关联标记
        in: query
        name: sysDictionaryID
        type: integer
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - description: 字典值
        in: query
        name: value
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用id查询SysDictionaryDetail
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询SysDictionaryDetail
      tags:
      - SysDictionaryDetail
  /sysDictionaryDetail/getDictionaryDetailsByParent:
    get:
      consumes:
      - application/json
      parameters:
      - description: 是否包含子级数据
        in: query
        name: includeChildren
        type: boolean
      - description: 父级字典详情ID，为空时获取顶级
        in: query
        name: parentID
        type: integer
      - description: 字典ID
        in: query
        name: sysDictionaryID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取字典详情列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/system.SysDictionaryDetail'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 根据父级ID获取字典详情
      tags:
      - SysDictionaryDetail
  /sysDictionaryDetail/getDictionaryPath:
    get:
      consumes:
      - application/json
      parameters:
      - description: 字典详情ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取字典详情路径
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/system.SysDictionaryDetail'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取字典详情的完整路径
      tags:
      - SysDictionaryDetail
  /sysDictionaryDetail/getDictionaryTreeList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 字典ID
        in: query
        name: sysDictionaryID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取字典详情树形结构
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/system.SysDictionaryDetail'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取字典详情树形结构
      tags:
      - SysDictionaryDetail
  /sysDictionaryDetail/getDictionaryTreeListByType:
    get:
      consumes:
      - application/json
      parameters:
      - description: 字典类型
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取字典详情树形结构
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/system.SysDictionaryDetail'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 根据字典类型获取字典详情树形结构
      tags:
      - SysDictionaryDetail
  /sysDictionaryDetail/getSysDictionaryDetailList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 禁用状态，根据status字段动态计算
        in: query
        name: disabled
        type: boolean
      - description: 扩展值
        in: query
        name: extend
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 展示值
        in: query
        name: label
        type: string
      - description: 层级深度，用于查询指定层级的数据
        in: query
        name: level
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 父级字典详情ID，用于查询指定父级下的子项
        in: query
        name: parentID
        type: integer
      - description: 层级路径，如 "1,2,3"
        in: query
        name: path
        type: string
      - description: 排序标记
        in: query
        name: sort
        type: integer
      - description: 启用状态
        in: query
        name: status
        type: boolean
      - description: 关联标记
        in: query
        name: sysDictionaryID
        type: integer
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - description: 字典值
        in: query
        name: value
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取SysDictionaryDetail列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取SysDictionaryDetail列表
      tags:
      - SysDictionaryDetail
  /sysDictionaryDetail/updateSysDictionaryDetail:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新SysDictionaryDetail
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysDictionaryDetail'
      produces:
      - application/json
      responses:
        "200":
          description: 更新SysDictionaryDetail
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新SysDictionaryDetail
      tags:
      - SysDictionaryDetail
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...